// src/app/core/models/project-details.ts
export class ProjectDetailsModel {
  name:         string;
  description:  string;
  language:     string;
  _id:          string;
  facebook:     string;
  twitter:      string;
  updated:      string;
  created:      string;
  intents:      number;
  dialog_nodes: number;
  entities:     number;
  edit:         boolean;
}